backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyPDF2==3.0.1
python-dotenv==1.0.0
snowflake-connector-python==3.4.0
pydantic==2.5.0
python-multipart==0.0.6
